Fix typos in the date-reading and writing.
I didn't test this codepath locally, I should have before.

I also fixed the code to not update the usage-file in dry-run mode.

I had to fix fetch_stats.py similarly as well, in particular not to
update the last-processed-time when no records were updated in a run.

Auditors: chris

Test Plan:
Ran
   echo "2015-07-20" > ~/dashboard_usage_date.db
   ./fetch_usage.py -n
and saw it start processing on 7/21.

I then ran
   date +%s -d '10 minutes ago' > ~/dashboard_report_time.db
   ./fetch_stats.py -nv
and saw it say:
   [...]
   Would update last-processed-time from 1438194239 to 1438194537
   [...]
   Would update last-processed-time from 1438194239 to 1438194710