Replace all execute() commands with a new execute_with_retries().
When running this in prod, I see periodic errors like:
   httplib.BadStatusLine: ''
This is a transient network error, most probably, and retrying is the
right thing to do.

Auditors: chris

Test Plan:
   env PATH=/home/csilvers/khan/webapp/third_party/frankenserver:$PATH ./fetch_stats.py -nv
and it emitted a bunch of lines with non-zero data in them.  Likewise
I ran
   env PATH=/home/csilvers/khan/webapp/third_party/frankenserver:$PATH ./fetch_usage.py -nv

There were no network errors in that time probably, but fetch_usage
returns a 404 in the normal course of operation, and the retry-logic
correctly didn't try to retry in that situation (and correctly
reraised the 404), so at least the common case wasn't affected.