Don't try to delete+add a phabricator repo when a github repo goes private.
(or goes public)

For historical reasons, we use https github urls for public repos and
ssh github urls for private repos.  When a repo changes from public to
private, the url changes and this script thinks that means we have a
new repository.

I changed the script so it doesn't think that anymore.

We probably should just move to having ssh urls for everything.  I
don't see any reason not to do that.  But a task for another day.

Auditors: chris

Test Plan:
Ran
   ./update_phabricator_repositories.py -n /tmp
on my local machine, and it said not to add or delete anything, even
though I had just made webapp-i18n-bigfile public.