Reduce the number of graphite_bridge errors via retrying and quieting.
Summary:
We now retrying the urlopen for downloading graphite data, just like
for downloading other things.  Boy we have a lot of retry functions
in the world!

I also quiet an error that is not actually a problem: uploading the
same data twice.  cloudmonitoring ignores the second time, so we don't
really care that it's happening.

There are still errors we don't try to deal with here, notably
timeouts, but this at least partially addresses
https://app.asana.com/0/25429045350268/64223318100954

Test Plan:
Fingers crossed -- will see how the next graphite-bridge run goes
after it's deployed!

Reviewers: chris

Reviewed By: chris

Subscribers: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D23384