Get more information about when graphite-bridge times out.
We give graphite-bridge 4 minutes to run, which is usually plenty of
time but not always.  But I don't know where the hanging is -- on the
initial connect to cloudmonitoring, or on a particular stat, or what.
By adding a signal handler to dump the stacktrace when `timeout` sends
us the SIGTERM, we should get more information.

Helps with https://app.asana.com/0/25429045350268/64223318100954

Auditors: chris

Test plan:
Ran this in python:
    >>> import signal, traceback, sys
    >>> signal.signal(signal.SIGTERM, lambda sig, stack: (traceback.print_stack(stack), sys.exit(1)))
and then in my shell ran
    % killall python
and saw it say:
    File "<stdin>", line 1, in <lambda>
and then exit.