A utility for sending OOM and other data from bigquery to graphite.
Summary:
Most data that we log we also send to graphite at the time that we log
it, so this post-processing step isn't necessary.  But some log
messages aren't created by us -- for instance, OOM errors -- and the
only want to process those logs for graphite is to do it post-facto.
That is what this script does.

Oops, forgot the webapp prefix.

Test Plan:
Ran `./logs_to_graphite.py -n` manually on my local machine, and the
output seemed plausible.

Fingers crossed

Reviewers: benkraft, chris

Reviewed By: chris

Differential Revision: https://phabricator.khanacademy.org/D26198