Update script to work with the new repositories API.
Phabricator has been updated to use some new-fangled apis for creating
and querying repositories.  I had to modify the code to work with
them.

The phabricator library we're using doesn't deal well with the
new-fangled APIs, so I had to update that code as well; I update the
substate here to use it.

Auditors: chris

Test Plan:
I ran
   ./update_phabricator_repositories.py -n /tmp
and saw that it said
   Adding 0 new repositories to phabricator
   Removing 0 deleted repositories from phabricator

This means the querying code, at least, is working properly, since it
found all the repos on github and nothing extra.

The new-repo-creating is still a big question mark.  I ran some of the
commands manually and they seemed to work, but the big test will be
when someone adds a new repo.