Revert "Upgrade oauth2client and google-python-api-client dependencies"
Summary:
This reverts commit 770d554817a9afe58b62420d78f90c343bd9c781 (D28286)

We decided not to upgrade the version of oauth2client or
google-python-api-client in alertlib, because of the cost of upgrading
those libraries in webapp, where there are many uses. Since alertlib is
deployed to the same machine as internal webserver, we need to downgrade
these dependencies here as well.

Test Plan:
Same as in D28286, ran `./fetch_usage -vn` and ensured it correctly runs.

We never deployed the old version to toby, so we don't need to downgrade
dependencies there.

Reviewers: alexanderforsyth, csilvers

Reviewed By: csilvers

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D28361