Use end_time instead of start_time to filter requests.
We now look through the streaming logs to find all requests that
*ended* between start_time_t and start_time_t + time_interval_seconds.
("start_time_t" is a bit of a confusing name).  We want "ended"
because that's when a request gets written to the logs.  As an
example, if a request took 10 minutes to run, if we were filtering on
start_time we'd miss it unless we happened to be running this script
on data 10 minutes in the past.  Usually we run it on data 1 minute in
the past, so we'd miss it.

Auditors: alexanderforsyth

Test plan:
Ran ./logs_bridge.py.  Results loooked similar, but somewhat
different, to before.