Make sure label values are always strings.
The `bq` binary is annoying in that the output it gives back is always
strings, even if the value in bigquery itself is an integer.  bq_io
attempts to deal with this by translating everything number-like that
it sees into a numeric type.

This mostly works well, but can fail for weird input users give us,
like the uesr who claimed their browser was `8`.  bq_io turned this
into a number, and stackdriver complained since it wants
browser-labels to be strings.

So now I just hard-convert every label to a string.  I think that's
the only place I need to worry about types being wrong here.

Auditors: alexanderforsyth, dhruv

Test plan:
Fingers crossed