Fetching GCE instance stats via cloud compute API and sending to stackdriver.
fetch_instance_stats is a script to fetch GCE instance stats (currently just the number of failed instances) via the cloud compute API. These statistics are sent to Stackdriver. It is expected that this script will be run periodically via a cron job.

Test Plan:
fetch_instance_stats_test.py was run and passed. This tests some of the individual functions in fetch_instance_stats as well as the entire `main` function with the API calls mocked out.