Normalize all metrics by week over week and by number of requests.
Added in normalizing the rest of the requests (besides just conversions) by week over week and by number of requests. This does mean that the Stackdriver graph will look weird when the new metric values start being pushed in (the value ranges will shift from 500-5000 to 0.8 to 1.2 for example). This should be fine in terms of future alerting, but could be confusing when looking at the historical graph. One alternative is to rename the metrics to specify that they are normalized.

Test Plan:
None