Don't have the retries step on each other.
It takes a few seconds for a job to cancel, so when we retry right
away we often find it fails because the first job still isn't
canceled, and bigquery gets upset we're trying to create a new job
with the same name.

One option would be to wait for the cancel to finish, but I just give
the new job a unique name.

Auditors: alexanderforsyth

Test plan:
Fingers crossed