Use a random job-name rather than an md5sum.
It turns out that we can't reuse job-ids even after a job has
succeeded (presumably there's a time-out here, but it's not documented
what it might be).  So when I was trying to run the same query once an
hour, it was failing with the old md5 approach.  Use a random number
instead.

Auditors: alexanderforsyth

Test plan:
Fingers crossed