Use a sub-table to limit the extra scanning work bigquery needs to do.
Summary:
If I do a logs_bridge query by hand on bigquery, one of the things it
lets me do after the results are in is to see how the query-optimizer
chose to execute the query.  Looking at that, it was clear that it
executes each of the dozens of subqueries we have independently.
Since each of those subqueries iterates over many more rows of the
datastore than we actually need (for instance, all of logs_hourly or
even logs_daily for the 'time ago' data), that means we were
repeatedly doing a lot of extra work.

This change avoids that extra work by doing the query in two stages.
In stage one, we just select the logs-records we care about (that is,
those in the appropriate time range) and copy them to a temporary
table.  In the second stage, we do our dozens of subqueries over the
temporary table, which avoids doing that extra work.

The query-optimizer analysis also shows that bigquery is choosing to
execute the subqueries in serial rather than parallel -- if I'm
reading it right, which I may well not be -- so it may be possible to
have even bigger improvements by just making a bunch of bq calls in
parallel rather than having one giant bq.  But I'll leave that as a
TODO for now; this finishes in under a minute which is all we need.

Test Plan:
I ran `./logs_bridge.py -nvv` and saw this output:
```
[2016-07-14 12:35:30,860 DEBUG] Creating the temporary table for querying over

[2016-07-14 12:36:00,117 DEBUG] Done creating temporary table khan-academy:logs_streaming.logs_bridge_1468524810_2954
...
[2016-07-14 12:36:27,129 DEBUG] Sending to stackdriver: [{'metric': {'labels': {u'module_id': 'multithreaded'}, 'type': u'custom.googleapis.com/logs.status.400.week_over_week'}, 'points': [{'interval': {'endTime': '2016-07-14T19:34:30Z', 'startTime': '2016-07-14T19:34:30Z'}, 'value': {'doubleValue': 1.4951370086602531}}]}, ...]
```

The values look plausible to me.

Also: 57 seconds!  By comparison, the old code took 3x as long (170 seconds).

Reviewers: dhruv, alexanderforsyth

Reviewed By: alexanderforsyth

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D29011