Use a different dataset, that prod-read has write access to.
logs_bridge.py runs as prod-read, but now that we need to create a
temporary table it needs some write permissions.  (`prod-read` is a
somewhat misnamed user in this case.)  I manually created a small,
somewhat scarily named dataset that I gave prod-read access to, so it
could create all the temp-tables it needs in that dataset.  This
change tells the code to use this new dataset.

Auditors: alexanderforsyth

Test plan:
I manually ran
   bq --format=none --headless --project_id khan-academy mk --expiration 60 khan-academy:logs_streaming_tmp_analysis.logs_bridge_1468871941_1543
on toby with success.