Fix a test to use short-circuiting.
I figured out where the weird 'not a number' error message was coming
from!  It turns out `test` does not do short-circuiting, so it was
evaluating `$elapsed_time -gt 3600` even when `elapsed_time` was
empty.

I switched to using the shell `&&`, which does short-circuit, instaed
of test's `-a`.  Problem should be solved.

Auditors: alexanderforsyth

Test plan:
Ran
    sh -c '[ -n "" ] && [ "" -gt 3600 ]'
without error.  By comparison, I also ran
    sh -c '[ -n "" -a  "" -gt 3600 ]'
and got
    sh: 1: [: Illegal number:

which is what I was getting in cron-mail as well.