Add a test for cronjobs failing.
Summary:
We want to know if cronjobs have been failing.  For other taskqueues
we use the X_TASKQUEUE_RETRY header, in webapp, to tell us.  But for
cron, that header isn't useful since GAE never retries failed cron
jobs.  Instead, we have to look in bigquery, to see if the cronjob has
been returning a non-200 response.

It's ok if frequent (minutely) cronjobs fail periodically; presumably
the next run will make up for it.  So we do a test that tells us
whether or not //all// the cronjob-runs in a given time-range have
failed.  We can use that to alert if a cronjob is failing
consistently.  (Or if an infrequently running cronjob is failing even
once.)

Test Plan: I ran `./logs_bridge.py -nvv` without error.  Rest is fingers crossed.

Reviewers: dhruv, alexanderforsyth

Reviewed By: alexanderforsyth

Differential Revision: https://phabricator.khanacademy.org/D29297