Fix a bad bug involving window-size in our logs_bridge analysis.
We were hard-coding the window-size to 60 seconds in one place, even
though we were passing in 5 minutes in most other places.  This means
that, in effect, we were running the analysis every 5 minutes but only
analyzing about 1 minutes' worth of data each time.  (It was actually
typically around 2-3 minutes depending, because of the way the
bigquery table decorators are implemented.)

Auditors: dhruv, alexanderforsyth

Test plan:
Ran
   ./logs_bridge.py -nvv --window-seconds=300
and saw it say
   Creating the temporary table for querying over by running ...
   WHERE end_time >= 1469046121 and end_time < 1469046421
which is indeed a 300 second time interval.