Add script to email us about our uptime
Summary:
See the comments for the details of the computation; basically I defined
downtime as the fraction of the time when at least 0.25% of requests fail.  The
script will email `infrastructure-blackhole` our uptime for each day this week,
the change from last week, and the weekly and monthly average.  After landing,
I'll hook it up to a cronjob to run Monday morning (so we get a summary for the
previous week).

I started this back when we had the monitoring fixit, but didn't actually
finish it until now.  The last few weeks' stats are actually kinda interesting
-- this priming stuff has brought our uptime down noticeably.  I'm hoping this
report will help us have a sense of how some of these issues affect us.  There
are a lot more variations on the stats I'd like to add, but this should get us
started, so we can see if it's useful.

Test Plan:
* `python -m smtpd -n -c DebuggingServer localhost:1025` to run a local SMTP
  debug server
* run the script (pointed at `localhost:1025`)
* examine the HTML output, and use `putsmail.com` to send it to myself and make
  sure it looks pretty.

Reviewers: csilvers, dhruv, chris

Subscribers:

Differential Revision: https://phabricator.khanacademy.org/D30310