Attempt to filter metrics by 'app'
Summary:
This is a wip approach to allowing logs_bridge to filter metrics by
'app'. For many of our metrics, combining webapp logs with react-render
logs can lead to noisy week over week alerts with lots of false
positives.

This patch tries to filter out relevant log lines for each metric when
an 'app' is specified, but testing revealed that this approach won't
necessarily work. `num_requests` right now is calculated as the overall
number of requests across all modules, so we'll need to refactor that if
we want the normalization to work.

Test Plan:

I ran logs bridge locally as follows:

./logs_bridge.py -c ~/trash/test_config.json -n -vv

Where test_config is set up as:
[
    {
        "metricName": "logs.latency.week_over_week",
        "labels": ["route"],
        "apps": ["webapp"],
        "num_unique_labels": 5,
        "unique_labels_sorting_field": "num_requests_by_field",
        "query": "AVG(latency)",
        "normalizeByDaysAgo": 7,
        "normalizeByRequests": true
    },
    {
        "metricName": "logs.react-render.latency.week_over_week",
        "labels": ["route"],
        "apps": ["react-render"],
        "num_unique_labels": 5,
        "unique_labels_sorting_field": "num_requests_by_field",
        "query": "AVG(latency)",
        "normalizeByDaysAgo": 7,
        "normalizeByRequests": true
    }
]

Then I looked at the debug output to see whether the num and
num_requests fields were set correctly, before they would be sent to
stackdriver.