Change the `route` label to include the module along with the url-route.
Summary:
The same url-route on different modules can do very different things!
In particular: `main:/` is very different on the default module than
on the react-render-server module!  The same could be true for other
routes as well.

We do not change the alert names, so suddenly there will be a bunch of
semantically duplicated labels (one with the module and one without)
on some of our alerts.  If that's a problem we may need to change the
names of some of those alerts to start them over.

To implement this, I gave more flexibility in general for how we map
from bigquery fields to stackdriver labels.

I am hoping this helps with
https://app.asana.com/0/198245187747885/198245187747904

Test Plan:
I ran
   ./logs_bridge.py -n
and saw it have lines like this:
   [2016-11-08 14:00:49,809 INFO] Collecting to send to stackdriver: logs.cronjobs.failures ({u'route': 'batch:main:/admin/auto_dumble/check_cb_session_endpoints [__cron]'}) 0.0 1478642355
showing that the route now includes the module-name.

I also saw:
   [2016-11-08 14:00:49,808 INFO] Collecting to send to stackdriver: logs.latency.week_over_week ({u'route': 'default:api.main:/api/internal/sat/record_activity [POST]'}) 2.89916862409 1478642355
showing that `(None)` gets mapped to `default` properly in this
context as well.

Finally, I saw:
   [2016-11-08 14:00:49,807 INFO] Collecting to send to stackdriver: logs.distinct_ips_by_language.week_over_week ({u'lang': 'pt'}) 1.09921640214 1478642355
   [2016-11-08 14:00:49,807 INFO] Collecting to send to stackdriver: logs.conversions.scratchpad_updated.week_over_week ({}) 1.16331375033 1478642355
showing that the code still works for other, more normal labels as
well, in addition to metrics that don't have labels at all.

Reviewers: colin, samantha

Reviewed By: colin, samantha

Subscribers: nabil, dhruv

Differential Revision: https://phabricator.khanacademy.org/D31898