Move to reporting for the top 9 browsers instead of the top 8.
We keep getting alerts that the week-over-week numbers for edge are
bad: 0.25 as much traffic this week as last:
https://app.google.stackdriver.com/metric-violation?violation=0.k3zyengwovh9

But analysis of the raw data doesn't bear that out.

What I believe is happening is this: our config only sends data from
the 8 most popular browsers.  And edge is neck and neck with some
other browser for spot #8.  Sometimes it wins and we send data for it,
which is probably around 1.0.  Sometimes it doesn't win and we send no
data, which stackdriver treats as 0.0.  Stackdriver does some
averaging to get 0.25, and boom alertsville!

The easy solution is to just use the 9 most popular browsers; number
10, AppleMail, is at half the traffic of #9 so the top 9 is likely to
be stable.

Obviously, this is not a long-term solution.  That will probably
involve understanding better how stackdriver does averaging, and not
changing our list of labels in a way that messes with it.

Hopefully resolves https://app.asana.com/0/198245187747885/208109945031691

Auditors: dhruv, nabil

Test plan:
Fingers crossed