Add an optional `project` argument to `bq_util.query_bigquery`.
Summary:
As part of project Cha Cha Ching, we will need to be able to query different
bigquery projects. This is most conveniently done with `bq_util.query_bigquery`,
which previously did not allow for that possibility. This commit adds this
optional argument with the same default value that was part of `bq_util.call_bq`.

Relatedly, to avoid redundancy, the formerly optional `project` argument
of `call_bq` is now mandatory. The sole existing call which did not pass an
explicit `project` has been updated to pass the former default argument, except
in `query_bigquery`, which passes along its new `project` argument to `call_bq`.

Test Plan: none

Reviewers: chris, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D32755