Add option to send historical data to fetch_gcp_billing.py.
Summary:
Many of the costs we are interested in monitoring for project Cha Cha Ching
are only reported to bigquery after a delay of a day or two.
This commit disallows the earlier use of a manually set start-time,
and instead always sets the start-time to midnight of the same UTC day.

Now, end-time defaults to 11:59:59PM of the current UTC day,
or the current time in the common case that that will be in the future.

When fetching old data, it will be timestamped with the given end-time,
expected to be 23:59:59UTC. This is in contrast to the data retrieved
on an hourly basis throughout the day, which will be timestamped with
the end-time corresponding to the moment of time when the script was invoked.

This should avoid writing the same moment of time twice since the hourly script
does not run on the 59 or zero minute.

Test Plan:
`./fetch_gcp_billing.py --dry-run --end-time`
`./fetch_gcp_billing.py --dry-run --end-time 2017-01-07`
`./fetch_gcp_billing.py --dry-run --end-time 2017-01-09`
`./fetch_gcp_billing.py --dry-run --end-time 2017-01-11`
all showed the values I would expect from cross-checking the bigquery web UI.

I also added print statements in main to check that start_time_t and end_time_t
were the values I expected.

Reviewers: chris, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D32962