Fix hourly dip in stats.
Summary:
Logs bridge experiences a dip in stats around the hour boundary.

https://app.asana.com/0/198245187747885/246191962447328/f

The problem seems to be that close to hour boundaries the tables for
previous stats are chosen incorrectly. For example, if the window is
5 minutes and the time is 4:59pm, then only the hourly table for 4pm
is queried, thus missing data from 5pm-5:04pm. This means that metrics
that compare current to week old stats will see a problem around the
weekly boundary because the week old stats will not have as many requests
as expected.

This problem also existed for daily boundaries.

There are also a few other changes in this diff.

- I made some changes suggested by Craig in D33204

- I added the option to specify the time window in logs_bridge_debug
  because the default for the script doesn't match the window that
  production logs bridge uses.

- I changed the temporary table to include request time information.
  I found this information useful for debugging.

Test Plan:
Try queries at the hour and day boundaries:

  logs_bridge_debug.py --window-seconds=300 --metric=logs.line_count.week_over_week --start_time `gdate -d "1:59pm today" +%s`

  logs_bridge_debug.py --window-seconds=300 --metric=logs.line_count.week_over_week --start_time `gdate -d "3:59pm today" +%s`

And make sure that it fixes the log spike at those times. (I'm assuming PST
3:59pm PST = 11:59pm UTC)

Reviewers: dhruv, csilvers

Reviewed By: dhruv, csilvers

Differential Revision: https://phabricator.khanacademy.org/D33221