Correctly regenerate job_name in case of query failure
Summary:
In D33204, I added an optional job_name param to allow for debugging of
past queries, but this caused problems when bigquery queries failed and
we attempted a retry.

This patch clears out the job name after each try, so a new job name
will be generated for each attempt. This means that if a user passes in
a job name and the first attempt fails, the passed in job_name will be
overwritten with a randomly generated one. This is perhaps not ideal,
but in practice we only use the passed in job_name to debug issues and
generally only when debugging interactively, so I predict we won't see
this be a big issue.

Test Plan:
In a python repl, I ran:
```
>>> import bq_util
>>> bq_util.query_bigquery('select * from dhruv')
```
to test an invalid query. Before this patch, the code would retry and
print out an error message of:

```
BQ call failed with this output: BigQuery error in query operation:
Already Exists: Job
khanacademy.org:deductive-jet-827:bq_util_6378104583621299433
```

After this patch, the error is the expected:
```
bq_util.BQException: -- Query failed after 2 retries: BigQuery error in
query operation: Error processing job
'khanacademy.org:deductive-jet-827:bq_util_401680771479397613': Table
name
cannot be resolved: dataset name is missing.
```

Reviewers: csilvers, amos

Reviewed By: csilvers, amos

Differential Revision: https://phabricator.khanacademy.org/D33231