Ignore the header row wen checking costs per row.
This email started triggering every day, about 3 days ago.  It turns
out it is because data[0] is the header row, and data[0][2] is `$`,
and `'$' > 750`.  Why this suddenly started happening a few days ago,
I have no idea.  It could just be that the mail was getting filtered
before, and changes to mailing lists recently (due to the reorg)
caused it to suddenly start working.  Shrug.

Since we sort by cost, I guess we could just check if `data[1][2] >
750`.  But the `any` makes the semantics clearer.

Auditors: benkraft

Test plan:
Fingers crossed -- we'll see what the mail looks like tomorrow.