Remove logs bridge config entries that cause a huge number of metrics to be sent to Stackdriver.
Summary:
It is an error to send more than 200 metrics to Stackdriver, although
there is currently no indication of this when testing locally.

TODO in a future commit (soon!): throw an error in dry-run mode if
there are more than 200 entries, since that causes an error in prod.

Test Plan:
`./logs_bridge.py --dry-run --verbose`
both exited successfully and sent a number of metrics to Stackdriver
well under the limit of 200, namely 112.

Reviewers: amos

Auditors: amos

Differential Revision: https://phabricator.khanacademy.org/D33679