Throw an error in dry-run mode if there are over 200 metrics.
Summary:
The stackdriver API accepts a maximum of 200 metrics in a request.
There was previously no indication of this issue in local testing.
Now an exception is raised in dry-run mode if the number of metrics
exceeds 200, and gives the user a descriptive error message.

Test Plan:
I added the following entry to logs_bridge.config.json:
```
    {
        "metricName": "logs.temp",
        "labels": ["device", "country", "browser"],
        "query": "NTH(91, QUANTILES(page_load_time, 101))"
    }
```
The number of permutations of these three labels easily exceeds the limit of
200 metrics. With this modification, `./logs_bridge.py --dry-run --verbose`
fails, as expected.

Without this modification to the config file, the same command succeeds,
as expected.

Reviewers: amos

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D33696