Change cutoff for bad client nav requests.
Summary:
Use slightly different values for client and server nav cutoffs.

Turns out that this makes only a small difference in the reported numbers.

The cutoffs are to exclude clearly bad data from the calculations. In
examining the data we have found that it's not uncommon for page loads
to be stalled for many minutes or even hours. Perhaps this is due to network
issues. It's not clear. It's most common in video pages which makes sense
since there may be network issues keeping the video from loading even
after the page loads.

This change reflects the fact that server nav requests are expected to take
longer than client nav.

Test Plan:
Manually tested metrics:

    python logs_bridge_debug.py --metric logs.page_load_time_by_country.pct90 --start_time `gdate -d "6pm 2017-03-01" +%s`
    python logs_bridge_debug.py --metric logs.page_load_time_by_page.pct90 --start_time `gdate -d "6pm 2017-03-01" +%s`

Reviewers: nabil

Reviewed By: nabil

Differential Revision: https://phabricator.khanacademy.org/D33938