Allow metrics to be sent by email.
Added an HTML template which we can use to format the metrics into a message.

Added pingdom to the template.

Added dependencies to requirements.txt

Changed how secrets are read to allow it to work when the repo isn't in the
expected place in the fs, which it isn't on my machine during dev work.

Test Plan:
Manually tested with:
email_reliability_metrics.py --dry-run

Auditors: nabil