Integrate graphs into email.
We should probably make pingdom and degraded service metrics persistent too.
Then we can add graphs for them.

I think we should redo the persistence format. I propose:

    {start_date: "YYYYMMDD",
     end_date: "YYYYMMDD",
     metrics: [
       {type: "TYPE OF METRIC",
        data: ...},
       ...
       ]
    },
    ...

Each record has a time and a list of metrics. Each metric has a type and some data appropriate for that metric.

This way we can more easily add metrics, and scan the db by date.

Here's an example metric:

    {"type: "page_load.by_page",
     "data": [{"page_load_time": 10.122, "page_load_page": "exercise_page"}, {"page_load_time": 5.731, "page_load_page": "curation_page"}, {"page_load_time": 5.547, "page_load_page": "scratchpad_page"}, {"page_load_time": 6.633, "page_load_page": "logged_out_homepage"}, {"page_load_time": 3.903, "page_load_page": "login_page"}, {"page_load_time": 3.968, "page_load_page": "signup_page"}, {"page_load_time": 11.095, "page_load_page": "video_page"}, {"page_load_time": 12.262, "page_load_page": "article_page"}, {"page_load_time": 9.797, "page_load_page": "logged_in_homepage"}]
    }

Also it would be nice to have a way to send test emails from a dev environment rather than having to deploy to toby to send an email.
This is especially true because I suspect that we'll want to do a bunch of minor tweaks to the email formatting.

Test Plan:
email_reliability_metrics.py --dry-run

Auditors: nabil