Save pingdom and service degradation data to a file. Related refactoring.
Summary:
This commit saves the uptime data from the Pingdom API and the service
degradation data from the email_uptime script to our _PAGE_LOAD_DATABASE.
This will make graphing either or both types of data easy, although this
commit does not add that functionality.

Relatedly, read from the same file to get the data to put into an email
template. Because we consider the page load performance graphs the most
critical part of the report, this was done in such a way that the report
can still go out if some pingdom or service degradation data is missing.

Test Plan:
`python email_reliability_metrics.py --dry-run`

Other than that, we will test by deploying to Toby, observing the email,
and tweaking based on that.

Reviewers: amos

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D34597