Add optional --start-time, --end-time and --save-graphs flags to email_reliability_metrics.
Summary:
Previously, certain internal functions allowed custom start and end times,
while others always used the defaults. This commit consistently allows
optional arguments for start and end time for all relevant functions in the module,
and adds the ability to pass unix times for these values at the command line.

This makes it easier to seed the database with past weeks of data.
It could prove similarly useful if we alter the data we report.

For similar convenience, add a flag to save graphs to the disk from the command line.

Test Plan:
I deleted my existing ~/email_reliability_metrics.db, and invoked the script
several times as follows:

```
./email_reliability_metrics --dry-run --start-time=`gdate -d "30 days ago" +%s` --end-time=`gdate -d "23 days ago" +%s`
./email_reliability_metrics --dry-run --start-time=`gdate -d "23 days ago" +%s` --end-time=`gdate -d "16 days ago" +%s`
./email_reliability_metrics --dry-run --start-time=`gdate -d "16 days ago" +%s` --end-time=`gdate -d "9 days ago" +%s`
./email_reliability_metrics --dry-run --save-graphs --start-time=`gdate -d "9 days ago" +%s` --end-time=`gdate -d "2 days ago" +%s`
```

This added four entries in my database, as intended.
Note that the fourth command also passes the flag to save the graphs to disk.

I examined the generated PNGs and verified that they looked as I expected,
graphing several weeks of data.

Reviewers: amos

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D34649