Add charts of precise weekly numbers to email_reliability_metrics. Base graphs from zero. Explain data sources.
Summary:
This commit enhances our email reliability metrics html template to include
html tables showing this week's numbers, in addition to the graphs.
These tables include a column of hardcoded baseline numbers.

The graphs have also been updated to always base the y-axis from zero.

Finally, we add some more explanation of the sources of the data to the email,
and make assorted stylistic changes.

Test Plan:
For convenience, I changed the line in send_email after `if dry_run:` to read:
`print html`

I then invoked the script with:

./email_reliability_metrics.py --dry-run --save-graphs --start-time=1490042503 --end-time=1490647306 > ~/test.html

I chose those particular dates because they were already contained in my local
database file, following my work in D34649. I will attach it for reference.
To reproduce this test, use the same file as your ~/email_reliability_metrics.db.
The start-time and end-time I passed are the results of, respectively:
```
gdate -d "9 days ago" +%s
gdate -d "2 days ago" +%s
```
Which are unix times sometime on March 20 2017 and March 27 2017.

The hardcoded baseline numbers are based on this same file.

After running the above command, I opened the html file and verified the table looked
as I intended, at least in a browser.
Using this approach, the graphs do not appear inline, but due to the use of the
--save-graphs flag, I was able to open the png graphs to verify their y-axes now
always start from zero.

To verify that the email looks how we intend in the inbox, I will test live on toby.

Reviewers: amos

Reviewed By: amos

Differential Revision: https://phabricator.khanacademy.org/D34662