Lint cleanup in email_bq_data.py
Summary:

This fixes some indentation issues in email_bq_data.py so that we can remove
the linter disable rule.  The main issue was that there was a gnuplot script
literal requiring string interpolation with a dict indented flush with the left
margin.  The required indetation of the dict contents was off from what looks
normal for python, so I wrapped the gnuplot script with `textwrap.dedent` and
indented it at the normal level of the function that contains it.

Test Plan:
- `gae_dashboard/email_bq_data.py --date 20170515 --dry-run`
- verify it completes successfully

Auditors: benkraft