Change nightly usage reports to account for requests split into multiple entries
Summary:
See https://app.asana.com/0/31965416896056/326589756326976

In our nightly reports of instance hours / memory / rpcs / etc. the route
`None` has suddenly been appearing at a much higher level.  It turns out the
main thing that's been going on here is that we've been seeing more frequent
occurrences where the logging API splits the logs for a single request into
multiple entries.  Since only one of these entries will have a KALOG line, then
this means that all the other logs for the same request will be bucketed into
the `None` route.

Additionally it turns out that when the logs are split, the latency field does
not report on the segment of time corresponding to the log lines in an entry,
but the entire time from the start of the request to when that log entry was
emitted.  This means that we were multiple counting time in the `None` bucket
as well when there were more than 2 log entries per request.

All of this can be fixed by deduplicating on request id before operating on the
logs.  When deduplicating, we get the value for a field using `FIRST` which
takes the first non-null value in a group.  In some queries, no change was
necessary, as they were filtering only down to fields based on KALOG, which
will only exist in one entry anyway.

I also incidentally changed `COUNT(*)` to `COUNT(1)` in a couple places.  This
doesn't actually make a difference in these queries, but this makes it clearer
that our intent is to count rows in the subquery so that future refactorers
won't have to worry about figuring out whether there are any nested/repeated
fields in scope in the inner query that would change the result.

Test Plan:
- hard code `dry_run` to `True` in the send_to_stackdriver and send_email
  functions so that we can test running the queries for real, but not actually
  send the reports
- `gae_dashboard/email_bq_data.py --date 20170515`
- verify that the results look largely the same as the actual reports for 5/15
  in the infrastructure-blackhole archive, except that the `None` route is much
  further down the list

Reviewers: benkraft, nabil

Reviewed By: benkraft, nabil

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D35735