Fix field_name to fetch bigquery cost numbers from bigquery.
Summary:
It appears google renamed this field within the last week.
I'm not sure if there is some announcement of these changes we
could use to make changes in advance rather than after jobs break.

This commit also includes minor whitespace changes to satisfy the linter,
which is more strict now than the last time this file was modified.

Test Plan:
`gae_dashboard/fetch_bigquery_cost.py --dry-run --verbose`

Before the change, this command failed with the same error as we see
in the toby-admin mailing list, involving the missing field.

After the change, the cost is fetched successfully.

Reviewers: csilvers, benkraft, colin, amos

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D36301